Moved string representations into a single module (#333)
Move string rendering of the modules in a separate module. This allows
us to share the repr with a jax version.

This is how it currently looks like as part of a PyTorch layer :)
```python
>>> create_planner("cartpole")
CartPolePlanner(
  (diff_mpc): AcadosDiffMpcTorch(
    N_horizon=5, nx=4, nu=1, casadi_type='SX'
    inputs:
      x0       (B, 4)     initial states
      u0       (B, 1)     fixates first-stage control (optional)
      params   dict       parameter overrides (see parameters)
    outputs:
      u_star   (B, 1)     optimal first-stage control (= u0 if given)
      x        (B, 6, 4)  state trajectory
      u        (B, 5, 1)    control trajectory
      value    (B, 1)      cost (V(x0), or Q(x0, u0) when u0 given)
    parameters:
      differentiable:
        name   splits  shape  default
        xref1  global  (1,)   [0.]
      non-differentiable:
        name   shape   default
        xref0  (6, 1)  [[0.], [0.], [0.], [0.], [0.], [0.]]
        xref2  (6, 1)  [[0.], [0.], [0.], [0.], [0.], [0.]]
        xref3  (6, 1)  [[0.], [0.], [0.], [0.], [0.], [0.]]
        uref   (6, 1)  [[0.], [0.], [0.], [0.], [0.], [0.]]
  )
)
```
