+++ leap_c/ocp/acados/parameters.py
@@ -1,4 +1,3 @@
-import re
 from dataclasses import dataclass, field
 from typing import TYPE_CHECKING, Any, Literal
 from warnings import warn
@@ -213,70 +212,10 @@ class AcadosParameterManager:
     def non_differentiable_symbols(self) -> ca.SX | ca.MX:
         return self._non_differentiable_parameter_store.get_symbols()
 
-    @staticmethod
-    def _format_array(arr: np.ndarray) -> str:
-        return re.sub(
-            r"\s+",
-            " ",
-            np.array2string(arr, max_line_width=np.inf, separator=", ", threshold=10, edgeitems=2),
-        )
-
-    def _repr_sections(self) -> str:
-        """Return the differentiable/non-differentiable section blocks (without the header).
-
-        Used by :meth:`__repr__` and embedded into ``AcadosDiffMpcTorch.__repr__``.
-        """
-        lines: list[str] = []
-
-        diff_names = self.differentiable_parameter_names
-        lines.append("  differentiable:")
-        if diff_names:
-            rows = []
-            for name in diff_names:
-                param = self.parameters[name]
-                splits = str(param.splits)
-                shape = str(param.overwrite_shape(self.N_horizon))
-                default = self._format_array(param.broadcasted_default(self.N_horizon))
-                rows.append((name, splits, shape, default))
-            w_name = max(len("name"), *(len(r[0]) for r in rows))
-            w_splits = max(len("splits"), *(len(r[1]) for r in rows))
-            w_shape = max(len("shape"), *(len(r[2]) for r in rows))
-            lines.append(
-                f"    {'name':<{w_name}}  {'splits':<{w_splits}}  {'shape':<{w_shape}}  default"
-            )
-            for name, splits, shape, default in rows:
-                lines.append(
-                    f"    {name:<{w_name}}  {splits:<{w_splits}}  {shape:<{w_shape}}  {default}"
-                )
-
-        nondiff_names = self.non_differentiable_parameter_names
-        lines.append("  non-differentiable:")
-        if nondiff_names:
-            rows = []
-            for name in nondiff_names:
-                param = self.parameters[name]
-                shape = (self.N_horizon + 1, *param.default.shape)
-                tiled = np.tile(param.default, (self.N_horizon + 1, *([1] * param.default.ndim)))
-                default = self._format_array(tiled)
-                rows.append((name, str(shape), default))
-            w_name = max(len("name"), *(len(r[0]) for r in rows))
-            w_shape = max(len("shape"), *(len(r[1]) for r in rows))
-            lines.append(f"    {'name':<{w_name}}  {'shape':<{w_shape}}  default")
-            for name, shape, default in rows:
-                lines.append(f"    {name:<{w_name}}  {shape:<{w_shape}}  {default}")
-
-        return "\n".join(lines)
-
     def __repr__(self) -> str:
-        df_size = self.differentiable_default_flat.size
-        ndf_size = self.non_differentiable_default_flat.size
-        header = (
-            f"AcadosParameterManager(N_horizon={self.N_horizon}, "
-            f"casadi_type='{self.casadi_type}', "
-            f"differentiable_flat={df_size}, "
-            f"non_differentiable_flat={ndf_size})"
-        )
-        return header + "\n" + self._repr_sections()
+        from leap_c.ocp.acados.repr import format_parameter_manager_repr
+
+        return format_parameter_manager_repr(self)
 
     @staticmethod
     def _create_symbol(name: str, size: int, casadi_type: Literal["SX", "MX"]) -> ca.SX | ca.MX:
