+++ leap_c/ocp/acados/torch.py
@@ -15,6 +15,10 @@ from leap_c.ocp.acados.diff_mpc import (
 )
 from leap_c.ocp.acados.initializer import AcadosDiffMpcInitializer
 from leap_c.ocp.acados.parameters import AcadosParameterManager
+from leap_c.ocp.acados.repr import (
+    format_diff_mpc_module_extra_repr,
+    format_diff_mpc_module_repr,
+)
 
 AcadosDiffMpcSensitivityOptions = Literal[
     "du0_dp_global",
@@ -190,35 +194,13 @@ class AcadosDiffMpcTorch(torch.nn.Module):
                 solver.constraints_set(stage, "ubx", ubx[i, j])
 
     def extra_repr(self) -> str:
-        ocp = self.diff_mpc_fun.ocp
-        N = ocp.solver_options.N_horizon
-        nx = ocp.dims.nx
-        nu = ocp.dims.nu
-        ct = self.parameter_manager.casadi_type
-        return f"N_horizon={N}, nx={nx}, nu={nu}, casadi_type='{ct}'"
+        return format_diff_mpc_module_extra_repr(
+            ocp=self.diff_mpc_fun.ocp, parameter_manager=self.parameter_manager
+        )
 
     def __repr__(self) -> str:
-        ocp = self.diff_mpc_fun.ocp
-        N = ocp.solver_options.N_horizon
-        nx = ocp.dims.nx
-        nu = ocp.dims.nu
-        sections = self.parameter_manager._repr_sections()
-        sections_indented = "\n".join(
-            "  " + line if line else line for line in sections.splitlines()
-        )
-        return (
-            "AcadosDiffMpcTorch(\n"
-            f"  {self.extra_repr()}\n"
-            "  inputs:\n"
-            f"    x0       (B, {nx})     initial states\n"
-            f"    u0       (B, {nu})     fixates first-stage control (optional)\n"
-            "    params   dict       parameter overrides (see parameters)\n"
-            "  outputs:\n"
-            f"    u_star   (B, {nu})     optimal first-stage control (= u0 if given)\n"
-            f"    x        (B, {N + 1}, {nx})  state trajectory\n"
-            f"    u        (B, {N}, {nu})    control trajectory\n"
-            "    value    (B, 1)      cost (V(x0), or Q(x0, u0) when u0 given)\n"
-            "  parameters:\n"
-            f"{sections_indented}\n"
-            ")"
+        return format_diff_mpc_module_repr(
+            class_name=type(self).__name__,
+            ocp=self.diff_mpc_fun.ocp,
+            parameter_manager=self.parameter_manager,
         )
